+++ src/main/java/karuna/karuna_backend/Controllers/PostController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import karuna.karuna_backend.DTO.PostCreateDto;
-import karuna.karuna_backend.DTO.PostDto;
+import karuna.karuna_backend.DTO.Post.PostCreateDto;
+import karuna.karuna_backend.DTO.Post.PostDto;
 import karuna.karuna_backend.Services.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
