+++ src/main/java/karuna/karuna_backend/DTO/Post/PostCreateDto.java
@@ -1,3 +1,3 @@
-package karuna.karuna_backend.DTO;
+package karuna.karuna_backend.DTO.Post;
 
 public record PostCreateDto(String body) { }
